decision_maker/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
langchain==0.2.7
langchain-openai==0.1.14
pydantic==2.7.4
python-dotenv==1.0.1
